requirements.txt
Django==4.0.4
kafka-python==2.0.2